// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/sprints-model.ts
export class SprintModel {
    sprintId: string;
    projectId: string;
    sprintName: string;
    projectName: string;
    sprintStartDate: any;
    sprintEndDate: any;
    createdDateTime: any;
    createdUserName: string;
    inActiveDateTime: Date;
    isBacklog: boolean;
    isArchived: boolean;
    timeStamp: any;
    pageNumber: number;
    pageSize: number;
    totalCount: number;
    description: string;
    isRefresh: boolean;
    isReplan: boolean;
    boardTypeId: string;
    boardTypeApiId: string;
    version: string;
    testSuiteId: string;
    isBugBoard: boolean;
    boardTypeUiId: string;
    workFlowId: string;
    isWarning: boolean;
    userStoriesCount: number;
    completedUserStoriesCount: number;
    sprintIds: string;
    sprintUniqueName: string;
    sprintResponsiblePersonName: string;
    profileImage: string;
    isComplete: boolean;
    allSprints: boolean;
    editSprint: boolean;
    isGoalsPage: boolean;
    sprintResponsiblePersonId: string;
    isEnableSprints: boolean;
    isEnableTestRepo: boolean;
    boardTypeName: string;
    isActiveSprints: boolean;
    isSuperAgileBoard: boolean;
    completedSprintDate: Date;
    sprintEndDatedeadLine: any;
    sprintEndDatetimeZoneOffSet: number;
    sprintStartDatetimeZoneOffSet: number;
    sprintStartDatedeadLine: any;
    sprintStatusIds: string;
    sprintResponsiblePersonIds: string;
    projectIds: string;
    ownerUserIds: string;
    userStoryStatusIds: string;
    deadLineDateFrom: string;
    deadLineDateTo: string;
    projectFeatureIds: string;
    bugPriorityIds: string;
    dependencyUserIds: string;
    bugCausedUserIds: string;
    userStoryTypeIds: string;
    versionName: string;
    userStoryName: string;
    workItemTags: string;
    includeArchived: any;
    deadLineDate: any;
    updatedDateFrom: any;
    updatedDateTo: any;
    createdDateFrom: any;
    createdDateTo: any;
    workItemsCount: number;
    isBacklogSprints: boolean;
    isReplanSprints: boolean;
    isCompletedSprints: boolean;
    isDeleteSprints: boolean;
}